% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.adcon}
\alias{process.adcon}
\title{Process ADCON data.}
\usage{
process.adcon(aws_dir, adt_dir)
}
\arguments{
\item{aws_dir}{full path to the directory containing the AWS_DATA folder on ADCON server.}

\item{adt_dir}{full path to the directory containing the AWS_DATA folder on ADT server.}
}
\description{
Get the data from ADCON database, parse and convert into ADT format.
}
