% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{replicator.tahmo}
\alias{replicator.tahmo}
\title{TAHMO replicator.}
\usage{
replicator.tahmo(aws_dir, adt_dir)
}
\arguments{
\item{aws_dir}{full path to the directory TAHMO containing the AWS_DATA folder.}

\item{adt_dir}{full path to the directory containing the AWS_DATA folder for ADT.}
}
\description{
Download data from TAHMO API.
}
