% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_failed_files.R
\name{upload.failed.files}
\alias{upload.failed.files}
\title{Re-upload files failed to upload.}
\usage{
upload.failed.files(awsNET, dirAWS, dirUP)
}
\arguments{
\item{awsNET}{name of the AWS network from \code{adt_network} table.}

\item{dirAWS}{full path to the directory containing the AWS_DATA folder.}

\item{dirUP}{full path to the directory containing the AWS_DATA folder on ADT server.}
}
\description{
Re-upload all files failed to upload from the previous task to ADT server.
}
